miniproject1/requirements.txt
flask==3.0.2
pymongo==4.6.1
python-dotenv==1.0.1
flask-login==0.6.3
kafka-python==2.0.2 